fix:bugs in import router ,fix fetch api call for other component
