+++ src/views/App.tsx
@@ -1,5 +1,5 @@
 import { Outlet, RouterProvider } from "react-router-dom";
-import router from "../Router/router";
+import router from "@/router/router";
 import { CoffeeProvider } from "@/Contexts/CoffeesContext";
 
 function App() {
