+++ src/views/Coffee's/Cappuccino.tsx
@@ -2,9 +2,7 @@ import { useCoffee } from "@/Contexts/CoffeesContext";
 import Cappu from "@public/Images/Cappu.jpg";
 //This Compo it might be updated So... tell me then change it -_-
 function Cappuccino() {
-  const { coffeeList } = useCoffee();
-
-  console.log(coffeeList);
+  const coffeeList = useCoffee();
 
   if (!coffeeList) {
     return <p>Loading coffee data...</p>;
@@ -19,7 +17,7 @@ function Cappuccino() {
 
   return (
     <div className="flex flex-col border-2 border-coffee-brown pb-14" style={{ backgroundColor: "wheat" }}>
-      <p className="text-center playfair">Discribe Coffee</p>
+      <p className="text-center playfair">Describe Coffee</p>
       <img src={Cappu} />
       <p className="text-center text-coffee-brown playfair text-2xl py-3 bold">{randomCoffee.name}</p>
       <p className="text-center playfair text-l">{randomCoffee.description}</p>
